Refactor Row
